2.1.12 Рефакторинг. Вынесет названия фильтров в файл с константами.
