+++ src/const.js
@@ -1,0 +1,8 @@
+const FilterName = {
+  ALL_MOVIES: `All movies`,
+  WATCHLIST: `Watchlist`,
+  HISTORY: `History`,
+  FAVORITES: `Favorites`
+};
+
+export {FilterName};
